Implement PreferSlowCiphers policy for cryptography compliance
This adds an enterprise policy which causes Chrome to prefer TLS 1.3
ciphers in a certain order which is required for compliance with a
specific cryptography compliance regime (CNSA).

The implementation (which plumbs the setting from //chrome through
//services/network to //net and down to BoringSSL) is very similar to
the PreferSlowKexAlgorithms policy, implemented in crrev.com/c/6951913.

The policy is marked experimental (future_on rather than supported_on)
because the configurations are still subject to change during
implementation and standardization.

Bug: 445967223, 437413977
Change-Id: I4068aad68f30a756a3d3854eb3031a6286585cdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6986002
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Commit-Queue: Lily Chen <chlily@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1521425}