oobe: resurrect network portal state check in version_updater
version_updater used to handle the case that the network is portalled by
reporting to its delegate so the Network Error / Portal UI is shown.

This was gated by `network_portal_detector::GetInstance()->IsEnabled())`
which started returning false unconditionally in CL:5458690 + CL:5716172

Re-enable this code.

A follow up is required to clean up browsertest usage of
NetworkPortalDetector::Enable() . That is tracked in
https://crbug.com/447558638 .

Bug: 447558645
Test: unit_test --gtest_filter=*VersionUpdater*
Test: Manual - join a portalled SSID during OOBE and wait for portal web UI to show up

Change-Id: I50c0642ebf83a54842d6637be8e77aa10928d86d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6983406
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Commit-Queue: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Renato Silva <rrsilva@google.com>
Cr-Commit-Position: refs/heads/main@{#1521446}