Using cached challenges during refresh
Instead of re-parsing the headers, use the existing mechanism for
processing challenge headers and look them up from the session. This is
only visible for sites in the case where they try to do multiple
challenges during registration, which is not technically allowed per the
spec.

Bug: 438783634
Change-Id: I478bc2d336033391f727dbe979b469e8755ce708
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6986901
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Reviewed-by: thefrog <thefrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1521449}