[AIAE] Change IsEligibleForBnpl to static
Reasons of this change are this function is more related to BNPL
functionality as a whole rather than to a particular instance of
BnplManager. Also by changing it to static, we don't need to instantiate
a BnplManager to invoke this helper method, which in some cases is very
difficult to do. Changing this to a static function could solve this
problem with the minimum cost.

Bug: 444684349
Change-Id: I51c703219a36704e249624646aae734074150ac4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6977814
Commit-Queue: James Sun <jamessun@google.com>
Reviewed-by: Vinny Persky <vinnypersky@google.com>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1521466}