[Uno-D] Integrate History sync experiment in Uno Phase 2
Migrated the `switches::kEnableHistorySyncOptinExpansionPill` to
`syncer::kReplaceSyncPromosWithSignInPromos` as the feature will now
be used as part of this launch.

Also inverted the flag priority between
`switches::IsAvatarSyncPromoFeatureEnabled()` and
`syncer::kReplaceSyncPromosWithSignInPromos` so that the latter has
priority when both are enabled (since sync would not be an option
anymore). Followup CLs will take care of the specific behavior for
Windows users.

Bug: 447048341
Fixed: 444617947, 418145883
Change-Id: Ica0434df370c2ca896c453a1c41dab6d91c0e291
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6983027
Commit-Queue: Ryan Sultanem <rsult@google.com>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1521497}