Reprocess challenge header after registration
To relieve site developers from having to understand the nuances of DBSC's
fetch integration, we want to allow the registration response to do both
registration and set a challenge.

This change is needed because normally, the challenge header is
processed before registration, and at that point there is no session to
cache the requested challenge onto.

Bug: 438783634
Change-Id: I0bd7fc43ee21ba18b0b9ef947085857bfc72269c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6990969
Reviewed-by: thefrog <thefrog@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1521515}