Use the same task runner for TFLite model loading and execution
This is a partial step towards moving all model loading onto the same task
runners as are used for execution, for the purpose of having identical
task execution priority.

Rationale: if running a model is important enough to have user visible
or user blocking impacts, it's very likely that clients will want to load
the model as soon as possible and not wait for best effort task execution.

Bug: 429938366
Change-Id: Ic6f582278e91a3e0cef1183211a9f97da560a033
Ax-Relnotes: n/a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6972851
Reviewed-by: Elias Klim <elklm@chromium.org>
Commit-Queue: Mike Wittman <wittman@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Auto-Submit: Mike Wittman <wittman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1521541}