ozone/drm: Propagate drm master drop result
For some reason, DrmGpuDisplayManager::RelinquishDisplayControl() does
not return any indication of success (unlike TakeDisplayControl() which
returns a bool). DrmThread just assumes the operation always succeeds
and returns a true for the success arg when invoking the callback.
Routine calls to drop master can fail, so the callback should be made
aware of the actual status of the call. There's already logic in
DisplayConfigurator to handle failure, and propagating this failure
will let the client know of the failure and let them retry (vs right
now where DisplayConfigurator thinks it doesn't have the master and
doesn't even try to drop it).

Bug: b:447219059
Test: New DrmGpuDisplayManager test
Change-Id: I911628fec558b81d43cf65b0e9fe33d6bb513ba5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6985701
Reviewed-by: Gil Dekel <gildekel@chromium.org>
Commit-Queue: Su Hong Koo <sukoo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1521564}