[History] Rename GetDailyVisitsToHost() to GetDailyVisitsToOrigin()
The method in VisitDatabase already constructs an origin from the
provided URL, and all leaf calls to the VisitDatabase method go through
HistoryService::GetDailyVisitsToOrigin, which takes a url::Origin. For
consistency, we should do the same in the lower level methods.

Encountered while preparing to add a new param to the HistoryService
method.

Bug: 441284623
Change-Id: Ifa8445a2692a60ad99faf4f8f65d65b90189cbb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6981589
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Commit-Queue: Svend Larsen <svend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1521577}