[Masonry] Use pre-existing ComputeExpandedPositions to assist in calculating individual track sizes for MasonryRunningPositions
This change uses the pre-exisiting ComputeExpandedPositions method to
assist in calculating and caching track sizes when dense-packing is
enabled. No tests were added as there should be no functional change.

This also adds the suggestions to move a conditional + fix a comment for
the dense-packing specified-span CL.

Bug: 343257585
Change-Id: I5bb18060e30f4ec8eaabc6a5dd373cf7ab847917
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6986061
Commit-Queue: Celeste Pan <celestepan@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1521578}