[VL] Add histogram for 404 VisitedLink partitioned hashtable entries
The histogram will be used to monitor the proportion of VisitedLink
partitioned hashtable additions that are added as a result of 404
visits, after enabling the saving of 404 visits to the History DB.

The histogram is emitted via macro rather than the corresponding
function because it is called in a loop for redirects, and
tools/metrics/histograms/README.md says: "when recording metrics on the
critical path (e.g. called in a loop or logged multiple times per
second), use the macros in histogram_macros.h instead [of the histogram
functions]."

This CL is part of a larger project to make links `:visited` after a 404
visit. Project links:
- Google-internal project doc: go/stop-filtering-history-by-status-code
- Issue tree / work status: https://crbug.com/427225439/dependencies

Fixed: 445899721
Change-Id: I449645f2e2bc2620f5bc52529aa0667b4b3e4e76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6975102
Reviewed-by: Kyra Seevers <kyraseevers@chromium.org>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Commit-Queue: Svend Larsen <svend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1521580}