Update WebrtcAudioSinkAdapter to support mono audio and fix a crash
The existing impl of this class was written in 2016 and was never
fully tested or implemented. When trying to get audio working for
the native RemotingClient to support the EDU team, I saw a large
number of warnings about an unsupported channel count. fast-
forward a bit after a few hours of digging and I understand them
well enough to make a fix.

The TL;DR is that WebRTC will deliver 'empty' audio samples to us
(i.e., with channels == 1) until the host *plays* audio, at that
point the audio will be decoded and I see the correct channel
number. I'm adding some code to handle mono or stereo for now
but we may decide to revert this and downgrade the warning to a
vlog depending on what the EDU folks find when they actually use
this class.

I also fixed a crash where ProcessAudioPacket passed an empty
OnceClosure rather than a DoNothing (which I think was the intent)
so I fixed that.

Bug: 443108152
Change-Id: I1ed23e5867f412bead3b2278c5646bb7ef0881fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6990105
Reviewed-by: April Zhou <aprilzhou@google.com>
Reviewed-by: Yuwei Huang <yuweih@chromium.org>
Auto-Submit: Joe Downing <joedow@chromium.org>
Commit-Queue: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1521690}