[pseudo&] Don't return selector from ExtractBestBucketingValues
ExtractBestBucketingValues currently returns the last simple selector
in the subject compound, so that the caller can look at the relation
(i.e. combinator) to the next CSSSelector in storage order [1],
and make some advanced bucketing decisions based on this.

Soon, however, pseudo-element selectors can appear within e.g. :is()
(Issue 444386484). At that point, there's not always a single useful
simple selector to return. For example, a selector can be both
shadow-crossing and and not shadow-crossing at the same time:

  :is(.a, #myhost::part(p))

Under the current "return last simple selector" behavior, the last
(top-level) simple selector here is the :is() itself, with relation
kSubSelector. That is not useful information for making the bucketing
decisions mentioned above.

No longer being able to classify a selector as shadow-crossing
*or* non-shadow-crossing is quite incompatible with how we gather
tree-crossing rules (we rely deeply on "accurate" bucketing),
and that is not something this CL attempt to address fully.
It's just a small step on the way.

Note: The existing bucketing code for part()/UA shadow pseudo-elements
explains that we should bucket by the *innermost* such pseudo-element.
However: I) A unit test expects the opposite [2]; II) the code appears
to only look at the next compound (non-recursively); III) we do not
actually allow (parse-time) any of the relevant combinations that code
is worried about except what is covered by [2] (the relevant branch
in the original CL:5851273 is shown as unexercised by tests);
IV) it's not clear from MatchOuterScopeRules why this bucketing
approach would be required; and V) we'll have to change much of
the behavior around shadow-crossing selectors for Issue 444386484
*anyway*, which will probably affect this bucketing behavior as well.
Therefore, this CL simplifies the behavior to "bucket according
to rightmost pseudo-element selector" for now.

Added some additional unit tests, especially for ::slotted(),
which was missing entirely. (The tests do not necessarily fail
without this CL.)

[1] See CSSSelector class comment.
[2] FindBestBucketAndAdd_ShadowPseudoAfterPart

Bug: 444386484
Change-Id: I9fe84510e9151a7fcb7f5b97eb7cd595a48402f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6969105
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1521718}