Disable PowerSaveBlocker on Linux.
When enabled, PowerSaveBlocker causes the host process to crash if the
DBus process exits. Although the host should exit if this happens (and
get restarted by the wrapper script), a crash is not ideal. This CL
disables PowerSaveBlocker until a more permanent solution can be found.

Change-Id: I63a49e3190cf5d22cfe99401e11a9788ba14c9c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6990180
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Lambros Lambrou <lambroslambrou@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1521778}