[VL] Let callers specify 404 policy in GetDailyVisitsToOrigin()
404 visits have historically not been eligible for History, so callers
of History querying methods have not needed to consider whether 404
visits are relevant for their feature or not. In the future, when
history::kVisitedLinksOn404 is enabled, 404s will become eligible for
History, and therefore be returnable by GetDailyVisitsToOrigin(). Not
all callers want to receive 404 visits, so this CL adds a
policy_for_404_visits param to the method to allow callers to specify
whether they want 404 visits.

This CL is part of a larger project to make links `:visited` after a 404
visit. Project links:
- Google-internal project doc: go/stop-filtering-history-by-status-code
- Issue tree / work status: https://crbug.com/427225439/dependencies

Fixed: 441284623
Change-Id: Ic7e35d7082aa4675c9b32f8dbd25ce223c9fbd0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6978869
Reviewed-by: Jian Li <jianli@chromium.org>
Reviewed-by: Tina Wang <tinazwang@chromium.org>
Commit-Queue: Svend Larsen <svend@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1521800}