AccountSettingSyncBridge skeleton
This CL creates a DataTypeSyncBridge for ACCOUNT_SETTING. All interface
methods are left NOTIMPLEMENTED(). The bridge is owned by the
AccoountSettingService keyed service and lives on the UI thread.
DataTypeStore will be used as the bridge's storage layer. Creation of
the bridge is guarded kSyncAccountSettings feature flag.

Bug: 441735283
Change-Id: Ifef69279ba1dc593aa47557efa1d500619345080
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6988933
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1521934}