Remove most RunUntilIdle() calls in HttpStreamFactoryControllerTests.
It was using a MockHttpStreamRequestDelegate that used gmock and
RunUntilIdle() to wait until it was called. This migrates
MockHttpStreamRequestDelegate off of gmock, in favor of giving it a
wait function and a RunLoop.

Some tests still use RunUntilIdle(), either to wait for other things
to happen (e.g., preconncts), or to make sure there are no pending
tasks that crash when run.

Bug: None
Change-Id: I2359cd78acf338bfe15a6b32aee342ab003e3e07
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6990910
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Commit-Queue: mmenke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1521995}