IDB: apply backing store verifier before/after in-session cleanup
Add functionality to verify (to some degree) that the contents of the
LevelDB backing store has not changed after an in-session cleanup has
completed. This functionality is controlled by a feature flag and is
intended to be run on a very limited number of real users (such as early
release channels).

This verification step requires reading the entire database twice and
keeping a snapshot in memory. Due to the performance implications of
that, it's limited to running on databases under a certain size and at
most on every 100th cleanup. The backing store snapshot code is updated
to compress larger databases more aggressively to avoid using too much
memory.

The motivation for this change is partially that a previous attempt to
roll out the in-session cleanup (tombstone sweeping) had to be halted
due to a destructive bug. If this verification step had been in place,
it would have alerted the existence of this bug earlier. Since this is
an operation that has shown to be error-prone, it makes sense to take
extra precautions.

The other motivation is that a similar process will likely be warranted
when rolling out the IndexedDB-on-SQLite data migration. We can run the
same or similar verification code then, so adding this functionality now
gives us some insight into how well this technique might work.

Bug: 404394638
Change-Id: I8f5240e4b9ce6d5345c5b25070e5e27ce4d9e7d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6955307
Commit-Queue: Abhishek Shanthkumar <abhishek.shanthkumar@microsoft.com>
Reviewed-by: Abhishek Shanthkumar <abhishek.shanthkumar@microsoft.com>
Reviewed-by: Mingyu Lei <leimy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1522016}