[indexeddb] Add IndexedDB.TransactionTimeout.10kTransactions.* histograms.
The existing IndexedDB.TransactionTimeout.* histograms were useful to
confirm that we frequently have timeouts with a lot (>10k) transactions
on the same connection, and that we frequently have timeouts with a
"commit pending" in the transaction. To verify whether these situations
occur at the same time (timeout with >10k transactions *and* commit
pending), this CL adds IndexedDB.TransactionTimeout.10kTransactions.*
histograms, which are the same as the existing
IndexedDB.TransactionTimeout.* histograms, but only recorded when there
are >10k transactions in the connection.

Bug: 381086791
Change-Id: I70db85bff4373ef2b797fb44919c4ea38c2bb735
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6974344
Commit-Queue: Mingyu Lei <leimy@chromium.org>
Auto-Submit: Francois Pierre Doray <fdoray@chromium.org>
Reviewed-by: Mingyu Lei <leimy@chromium.org>
Reviewed-by: Austin Sullivan <asully@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1522033}