[geo] Launch location permission subpage from page info
This change launches a subpage from the GEOLOCATION_WITH_OPTIONS
permission preference in page info to edit the permission precision.

This creates a new page info controller for the subpage and launches it
from the PermissionsController. A new SingleWebsiteSettings.Observer
method is added to notify the PermissionsSontroller when the subpage
should be opened.

Since the Website is not available in the controller, an
EXTRA_SITE_ADDRESS was introduced to the location subpage fragment to
fetch the Website in an alternate method.

Bug: 418936295
Change-Id: Id67216b72296057c1bedcaf2ca8b78066db5f89a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6988627
Commit-Queue: Zaina Al-Mashni <zalmashni@google.com>
Reviewed-by: Filipa Senra <fsenra@google.com>
Cr-Commit-Position: refs/heads/main@{#1522086}