[AccountName] Implement soft and hard removals
The kAccountNameEmail profile was previously permanently removed if the
user signed out. This is not ideal, as signing out is often temporary,
and the user likely expects the profile to reappear upon signing back
in.

This change introduces a distinction between a "soft" (temporary) and
"hard" (permanent) removal of the profile.

- A **soft remove** now occurs when a user signs out or disables
Autofill sync toggle. The profile is removed but will be restored if the
user signs back in or re-enables autofill syncing - this profile is only
available for the syncing users.

- A **hard remove** still occurs when a user explicitly deletes the
profile, repeatedly ignores its suggestions, or imports a superset
profile. This action prevents the profile from being automatically
recreated.

The profile will still be recreated after a hard removal if the extended
account info changes, as this is considered a new profile.

This CL was created as a result of splitting:
https://chromium-review.googlesource.com/c/chromium/src/+/6935443

Bug: 443047818
Change-Id: I0dcfc048e3ba5eeacc25fff886b66ecaf660c865
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6973536
Commit-Queue: Karol Sygiet <sygiet@google.com>
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1522113}