OTP field event logger for metrics.
This CL adds an event logger for OTP fields. This event logger should
generate key metrics identical to the ones use for Addresses and Cards
with a situation specific definition for Readiness.

Specifically, the logger uses the following definition of Readiness:
Chrome is ready for filling an OTP field if during the lifetime of the
form there was at least one OTP available to fill.

Note that there is a subtle difference to the Readiness definition of
addresses where the availability at filling time is checked. Since
OTP can expire - although unlikely to happen before submission - we
nevertheless count every OTP over the full lifetime of the form.


The test infrastructure has a flakiness issue on MacOs. This needs
careful investigation. Since we are not targeting MacOs, the tests are
currently disabled on MacOs to allow landing tests, but understanding
and resolving the flakiness is an important next step.

Change-Id: I05d5b248b2fd352566a58ef1d3446a0be5441f6f
Fixed: 443693025, 415272525, 415272525
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6939388
Auto-Submit: Matthias Körber <koerber@google.com>
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Commit-Queue: Matthias Körber <koerber@google.com>
Cr-Commit-Position: refs/heads/main@{#1522119}