Add histograms for OSCrypt migration.
This CL adds histograms to track the success and failure rates of
encryption and decryption operations related to the sync bootstrap token
and Nigori storage. These metrics are needed for the upcoming Finch
experiment for migrating the encryption library in the sync component to
use OSCrypt.

Bug: 419157433

Change-Id: I29018f4c13799a8b0cf802e29bf06718ee075423
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6994758
Commit-Queue: Michael Tatarski <mtatarski@google.com>
Reviewed-by: Maksim Moskvitin <mmoskvitin@google.com>
Cr-Commit-Position: refs/heads/main@{#1522155}