[iOS] Add a content configuration for a switch view
This CL adds a content configuration for a switch view.
It also adds a trailing configuration for the TableViewCell content
configuration.

The switch target and selector are passed to the configuration to avoid
casting the cell and setting it here.

For now it is not using the cells as it would require rewriting the
EG helpers.

Before:
https://screenshot.googleplex.com/9kLGmBtctu5k5Hd
https://screenshot.googleplex.com/9X54X3QpkTSyqxC

After:
https://screenshot.googleplex.com/BsXkxwhX9Wciwoh
https://screenshot.googleplex.com/5raKbcuXxYJxnxd

Bug: 443034511
Change-Id: Ia2d246a4b277ac5d886070b788b9db0f687d5808
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6988187
Reviewed-by: Louis Romero <lpromero@google.com>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Louis Romero <lpromero@google.com>
Cr-Commit-Position: refs/heads/main@{#1522163}